backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
dj-database-url
gunicorn
psycopg2-binary
python-dotenv==1.0.0
mysqlclient
